requirements.txt
anyio==3.6.2
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==5.5
dj-database-url==1.2.0
Django==4.0.10
django-cors-headers==3.14.0
django-filter==23.1
django-safedelete==1.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Flask==2.2.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
Pillow==9.5.0
psycopg2-binary==2.9.5
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyJWT
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-telegram-bot==20.2
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.4
reportlab==3.6.12
requests==2.28.2
rfc3986==1.5.0
selenium==3.141.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
whitenoise==5.3.0
wikipedia==1.4.0
xhtml2pdf==0.2.10
boto3==1.17.53
botocore==1.20.53
